(#321) Show sub-titles of the header and footer in the navigation panel
All level-2 titles located in the header and the footer files and that
have an id in the /api-group-something/ format will get an entry
in the navigation bar.

Examples:

Here we use a <span> tag after the ## Markdown tag. The group is "intro"
and the rest is "start-well":

  ## <span id="api-intro-start-well">Starting Well with APIDoc</span>

Same example, using the alternate ----- Markdown tag for level 2 titles:

  <span id="api-intro-start-well">Starting Well with APIDoc</span>
  ----------------------------------------------------------------

Same example, using HTML instead of Markdown:

  <h2 id="api-intro-start-well">Starting Well with APIDoc</h2>

Remember that it is important to have your id start by api-, then have another
word (the group), a dash and a name (which can contain dashes, that's OK).

After that, in the rest of your documentation, you can easily point to these
entries using regular Markdown:

  Learn [how to start well with APIDoc](#api-intro-start-well)