Allow parsing of json parameter fields
based on the given parameter specification for those fields. Values are
entered in form as string values. Adding an extra step before sending
json body to parse Boolean and Number parameters from string so that the
final json sent in the request body is valid according to the parameter
specification.

Used _.get and _.set functions to work with nested paths as well.