Send sample request auto json body if body/json is not used but Content-Type application/json header is set (#862)
* Allow setting json body automatically

from parameters if Content-Type application/json header is given and
no explicit body/json has been set.

* Allow 'optional' parameter replacement

* Allow parsing of json parameter fields

based on the given parameter specification for those fields. Values are
entered in form as string values. Adding an extra step before sending
json body to parse Boolean and Number parameters from string so that the
final json sent in the request body is valid according to the parameter
specification.

Used _.get and _.set functions to work with nested paths as well.

* Add necessary argument

* Update lodash to include _.get and _.set

* Add _.isNaN