Make the 'optional' label unselectable
This seems to affect only some browsers (happens in Chrome, not in Firefox). When you double click a parameter name that is optional, the browser also selects the word 'optional'. This is pretty annoying when you want to double click and copy a property name, as the actually copied value becomes "propertyNameoptional".

Assuming that no one actually needs to select the text inside the optional label, I figured this was the simplest solution to the problem. Open to other solutions.