Add anchor tag links to each article group and name
Previously, if you wanted to link to a specific article, you needed to find it in the menu first. With a large API documentation, this was cumbersome. This change allows the user to click a link next to a group or article which adds the anchor tag to the URL.

I'm not sure this is a nice way to structure the HTML, but at least it didn't cause any line breaks or require any changes to CSS.

As part of this change I needed to upgrade bootstrap, as the old version was loading bootstrap from a non-https CDN, which Chrome blocks. I have compared the sample site using the old and new bootstrap version, and found nothing that broke.